Allow a unique keyword to be assigned to a tab.
This was inspired by keyword bookmarks, which allow you to type a keyword into the address bar.

To assign a keyword to the current tab, you type the = character into the text box, then type the keyword to assign and press enter.
If you then later type this exact keyword into the text box, this tab will show at the top of the results such that pressing enter will switch to this tab.
Note that the keyword must be an exact match.
This allows you to very quickly switch to frequently used tabs.